Fix role name comparisons (toLowerCase)
Call .toLowerCase() on the roleName being compared, as only one side was
guaranteed to be in lower case before this. The non lower case comparison led to the muted role being created every time you started up the bot if it was "Muted" instead of "muted", for example. The permissions override was also attempted every time due to this.