fix alias bug introduced with services port with a hack
Alias commands need to be loaded after the commands are registered with
the command container, which is after the configure block. Since KUtils
services are only injected to commands / services / listeners and are
privately held otherwise, there is no way to get a reference to the
service to load aliases at the right time.

The hack stores a reference in a companion object which allows us to get
a reference in the startBot block. This is temporary solution until a
new version of KUtils can be released which fixes this.

(cherry picked from commit ee4f7b5e9ad31a1f6eb7b1c6e818c5aaef1666d2)